+++ NEWS.md
@@ -1,6 +1,9 @@
 Version 0.17.0
 2014-MM-DD
 
+ - Use a single cache of localized resources, purge after
+   1 day of no access (#82)
+
 Version 0.16.0
 2014-01-14